% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_wrapper.R
\name{formula_wrap}
\alias{formula_wrap}
\title{Wrap a model function to accept a formula interface}
\usage{
formula_wrap(model_func, x_name = "x", y_name = "y")
}
\arguments{
\item{model_func}{is the model function to wrap. It should take a predictor matrix and response vector as arguments.}

\item{x_name}{is the name of the predictor matrix in the model function.}

\item{y_name}{is the name of the response vector in the model function.}
}
\value{
a new function that accepts a data frame and a formula, and calls the original model function with the appropriate arguments.
}
\description{
Wrap a model function to accept a formula interface
}
\examples{
\dontrun{
formula_glmnet <- formula_wrap(glmnet)
formula_glmnet(mtcars, mpg ~ cyl + hp)
}

}
