% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine.R
\name{mine}
\alias{mine}
\title{Title}
\usage{
mine(
  data,
  response_var,
  model_func = lm,
  max_degree = 3,
  max_interact_vars = 2,
  metric = AIC,
  metric_comparison = min,
  keep_all_vars = FALSE
)
}
\arguments{
\item{metric_comparison}{is a function that compares the two metrics and returns the preferable one. Defaults to min, can be max or some other comparison function.}

\item{keep_all_vars}{if set to TRUE, will keep all of the first-order terms in the model, just testing interactions and higher-order effects.}
}
\description{
Title
}
